Refactored the discovery state out into it's own file
